+++ angel_orm_generator/test/standalone_test.dart
@@ -12,24 +12,16 @@ main() {
   test('to where', () {
     var query = new CarQuery();
     query.where
-      ..familyFriendly.equals(true)
+      ..familyFriendly.isTrue
       ..recalledAt.lessThanOrEqualTo(y2k, includeTime: false);
     var whereClause = query.where.compile(tableName: 'cars');
     print('Where clause: $whereClause');
     expect(whereClause,
-        'WHERE cars.family_friendly = TRUE AND cars.recalled_at <= \'2000-01-01\'');
+        'cars.family_friendly = TRUE AND cars.recalled_at <= \'2000-01-01\'');
   });
 
   test('parseRow', () {
-    var row = [
-      0,
-      'Mazda',
-      'CX9',
-      true,
-      dateYmdHms.format(y2k),
-      dateYmdHms.format(y2k),
-      dateYmdHms.format(y2k)
-    ];
+    var row = [0, 'Mazda', 'CX9', true, y2k, y2k, y2k];
     print(row);
     var car = new CarQuery().deserialize(row);
     print(car.toJson());
@@ -59,20 +51,22 @@ main() {
         Car ferrari;
 
         setUp(() async {
-          ferrari = await CarQuery.insert(connection,
-              make: 'Ferrari',
-              description: 'Vroom vroom!',
-              familyFriendly: false);
+          var query = new CarQuery();
+          query.values
+            ..make = 'Ferrari東'
+            ..description = 'Vroom vroom!'
+            ..familyFriendly = false;
+          ferrari = await query.insert(connection);
         });
 
         tearDown(() => connection.close());
 
         test('where clause is applied', () async {
-          var query = new CarQuery()..where.familyFriendly.equals(true);
+          var query = new CarQuery()..where.familyFriendly.isTrue;
           var cars = await query.get(connection);
           expect(cars, isEmpty);
 
-          var sportsCars = new CarQuery()..where.familyFriendly.notEquals(true);
+          var sportsCars = new CarQuery()..where.familyFriendly.isFalse;
           cars = await sportsCars.get(connection);
           print(cars.map((c) => c.toJson()));
 
@@ -85,22 +79,20 @@ main() {
 
         test('union', () async {
           var query1 = new CarQuery()..where.make.like('%Fer%');
-          var query2 = new CarQuery()..where.familyFriendly.equals(true);
+          var query2 = new CarQuery()..where.familyFriendly.isTrue;
           var query3 = new CarQuery()..where.description.equals('Submarine');
-          query1
-            ..union(query2)
-            ..unionAll(query3);
-          print(query1.compile());
-
-          var cars = await query1.get(connection);
+          var union = query1.union(query2).unionAll(query3);
+          print(union.compile());
+          var cars = await union.get(connection);
           expect(cars, hasLength(1));
         });
 
         test('or clause', () async {
           var query = new CarQuery()
             ..where.make.like('Fer%')
-            ..orWhere((where) =>
-                where..familyFriendly.equals(true)..make.equals('Honda'));
+            ..orWhere((where) => where
+              ..familyFriendly.isTrue
+              ..make.equals('Honda'));
           print(query.compile());
           var cars = await query.get(connection);
           expect(cars, hasLength(1));
@@ -132,25 +124,28 @@ main() {
 
         test('delete stream', () async {
           var query = new CarQuery()
-            ..where.make.equals('Ferrari')
-            ..orWhere((w) => w.familyFriendly.equals(true));
-
+            ..where.make.equals('Ferrari東')
+            ..orWhere((w) => w.familyFriendly.isTrue);
           print(query.compile(preamble: 'DELETE FROM "cars"'));
-          var cars = await query.get(connection);
+
+          var cars = await query.delete(connection);
           expect(cars, hasLength(1));
           expect(cars.first.toJson(), ferrari.toJson());
         });
 
         test('update', () async {
-          var query = new CarQuery()..where.id.equals(int.parse(ferrari.id));
-          var cars = await query.update(connection, make: 'Hyundai');
+          var query = new CarQuery()
+            ..where.id.equals(int.parse(ferrari.id))
+            ..values.make = 'Hyundai';
+          var cars = await query.update(connection);
           expect(cars, hasLength(1));
           expect(cars.first.make, 'Hyundai');
         });
 
         test('update car', () async {
           var cloned = ferrari.copyWith(make: 'Angel');
-          var car = await CarQuery.updateCar(connection, cloned);
+          var query = new CarQuery()..values.copyFrom(cloned);
+          var car = await query.updateOne(connection);
           print(car.toJson());
           expect(car.toJson(), cloned.toJson());
         });
@@ -159,11 +154,13 @@ main() {
 
     test('insert', () async {
       var recalledAt = new DateTime.now();
-      var car = await CarQuery.insert(connection,
-          make: 'Honda',
-          description: 'Hello',
-          familyFriendly: true,
-          recalledAt: recalledAt);
+      var query = new CarQuery();
+      query.values
+        ..make = 'Honda'
+        ..description = 'Hello'
+        ..familyFriendly = true
+        ..recalledAt = recalledAt;
+      var car = await query.insert(connection);
       expect(car.id, isNotNull);
       expect(car.make, 'Honda');
       expect(car.description, 'Hello');
@@ -179,7 +176,8 @@ main() {
           description: 'Herbie',
           familyFriendly: true,
           recalledAt: recalledAt);
-      var car = await CarQuery.insertCar(connection, beetle);
+      var query = new CarQuery()..values.copyFrom(beetle);
+      var car = await query.insert(connection);
       print(car.toJson());
       expect(car.make, beetle.make);
       expect(car.description, beetle.description);